Split gem into multiple modules and classes. Add support for Adapters.
* Works exactly the same as the previous versions, but I've moved some of
  the code into modules and classes for easier editing.

* Adds Adapters, which allows the gem to use different HTTP libraries
  (see bottom of README.md for an example that uses open-uri).
