+++ lib/twitch/adapters/httparty_adapter.rb
@@ -1,0 +1,14 @@
+require 'httparty'
+
+module Twitch
+  module Adapters
+    class HTTPartyAdapter < BaseAdapter
+      
+      def self.request(method, url, options={})
+        res = HTTParty.send(method, url, options)
+        {:body => res, :response => res.code}
+      end
+
+    end
+  end
+end
