+++ lib/twitch/version.rb
@@ -1,0 +1,10 @@
+
+module Twitch
+  module VERSION
+    MAJOR = 0
+    MINOR = 0
+    TINY  = 5
+
+    STRING = [MAJOR, MINOR, TINY].compact.join(".")
+  end
+end
