+++ twitch.gemspec
@@ -1,6 +1,10 @@
+$:.push File.expand_path("../lib", __FILE__)
+
+require "twitch/version"
+
 Gem::Specification.new do |s|
   s.name        = 'twitch'
-  s.version     = '0.0.4'
+  s.version     = Twitch::VERSION::STRING
   s.date        = Date.today.to_s
   s.summary     = "Twitch API"
   s.description = "Simplify Twitch's API for Ruby"
@@ -8,7 +12,7 @@ Gem::Specification.new do |s|
   s.email       = 'dustin.lakin@gmail.com'
   s.homepage    = "https://github.com/dustinlakin/twitch-rb"
 
-  s.files       = ["lib/twitch.rb"]
+  s.files       = Dir["lib/**/*"]
   s.require_paths = ["lib"]
   
   s.add_dependency('httparty')
